--- src/main/java/community/whatever/onembackendjava/controller/request/ShortenUrlRequest.java
@@ -1,7 +1,0 @@
-package community.whatever.onembackendjava.controller.request;
-
-public record ShortenUrlRequest(
-    String originUrl
-) {
-
-}
