+++ src/main/java/community/whatever/onembackendjava/urlshorten/controller/UrlShortenController.java
@@ -1,9 +1,9 @@
-package community.whatever.onembackendjava.controller;
+package community.whatever.onembackendjava.urlshorten.controller;
 
-import community.whatever.onembackendjava.controller.response.GetOriginUrlResponse;
-import community.whatever.onembackendjava.controller.request.ShortenUrlRequest;
-import community.whatever.onembackendjava.controller.response.ShortenUrlResponse;
-import community.whatever.onembackendjava.service.UrlShortenService;
+import community.whatever.onembackendjava.urlshorten.dto.GetOriginUrlResponse;
+import community.whatever.onembackendjava.urlshorten.dto.ShortenUrlRequest;
+import community.whatever.onembackendjava.urlshorten.dto.ShortenUrlResponse;
+import community.whatever.onembackendjava.urlshorten.service.UrlShortenService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
