+++ src/main/java/community/whatever/onembackendjava/urlshorten/dto/GetOriginUrlResponse.java
@@ -1,4 +1,4 @@
-package community.whatever.onembackendjava.controller.response;
+package community.whatever.onembackendjava.urlshorten.dto;
 
 public record GetOriginUrlResponse (
     String originUrl
