+++ src/main/java/community/whatever/onembackendjava/urlshorten/dto/ShortenUrlRequest.java
@@ -1,0 +1,7 @@
+package community.whatever.onembackendjava.urlshorten.dto;
+
+public record ShortenUrlRequest(
+    String originUrl
+) {
+
+}
